refactor(file structure): move modules-chrome/* to modules/chrome/*
