+++ core/background/action.js
@@ -1,6 +1,6 @@
 import * as data from './pageInfo.js'
-import * as storage from '../../modules-chrome/storage.mjs'
-import * as tabs from '../../modules-chrome/tabs.mjs'
+import * as storage from '../../modules/chrome/storage.mjs'
+import * as tabs from '../../modules/chrome/tabs.mjs'
 import * as request from './request.js'
