+++ features/status-bar/statusBar.js
@@ -1,4 +1,4 @@
-import * as tabs from '../../modules-chrome/tabs.mjs'
+import * as tabs from '../../modules/chrome/tabs.mjs'
 import * as action from '../../modules/domActions.mjs'
 
 export function setup() {
