+++ modules/chrome/tabs.mjs
