+++ modules/domActions.mjs
@@ -1,4 +1,4 @@
-import * as extension from '../modules-chrome/runtime.mjs'
+import * as extension from './chrome/runtime.mjs'
 
 const activeLi = () => $('.active')
 const visibleLis = () => $('#reading-list li:visible')
